Create 18-9.py
Made the mistake of using array slicing to do the 'marble' circle.   This worked for part one, but was way to slow for part 2.   Tried to find some sort of pattern to take advantage of, but could not find any.   I ended up biting the bullet and completely re-coding the solution using a doubly linked list, costed me an hour, but worked in the end.