Create 18-10.py
Worked out roughly how to get how many iterations were required by dividing the position by the velocity.  Then assumed the answer would be somewhere in a 300 second window.  Assumes the message is the dataset that has the thinest spread.   When spread started getting bigger I rewound one and assumed that was the answer.