Create 18-22.py
Used a breadth first search (Dijkstra), i.e. iteratively at each second populate a  map of where we 'could' be located in the cave.  If we reach a place that we have been before terminate that line. Do this until the destination is found.
The mistake that costed me a lot of time was that when changing equipment, the new equipment must be satisfactory for both the start and destination squares, not just the destination, as I had assumed.
This solution should work for all cases.  This solution could be done using recursion also.