Create 19-7-pt2.py
Part 1 and 2 split in half, I figured it was too much work to merge.  Part 1 was quite different to Part 2, and much of the code had to be reworked to allow concurrent amps. 

Unfortunately my day 5 solution was messed up due to over cleaning which made it wrong, so caused problems for day 7.