Create 20-9.py
For the first part I got lazy and used n squared algorithm instead of an nlogn.  Worked out okay because it was a small puzzle.  For part 2 I figured the complexity would go up so I went for an nlogn for part 2, but I reckon I probably could have gotten away with an n squared algorithm. It's a trade off between execution time vs coding time. Coding time usually wins out for the early puzzles in aoc.  For the later puzzles it is have a go at a simple inefficient algorithm then when its taking too long press ctrl-C and then write a better algorithm.